+++ common/crypto/kerl/hash.h
@@ -5,18 +5,19 @@
  * Refer to the LICENSE file for licensing information
  */
 
+#ifndef __KERL_HASH_H
+#define __KERL_HASH_H
+
+#include "common/crypto/kerl/kerl.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#ifndef __KERL_HASH_H
-#define __KERL_HASH_H
-
-#include "common/kerl/kerl.h"
-
 void kerl_hash(const trit_t* const trits, size_t len, trit_t* out, Kerl* kerl);
 
-#endif  //__KERL_HASH_H
 #ifdef __cplusplus
 }
 #endif
+
+#endif  //__KERL_HASH_H
