+++ common/crypto/kerl/kerl.c
@@ -7,8 +7,8 @@
 
 #include <assert.h>
 
-#include "common/kerl/converter.h"
-#include "common/kerl/kerl.h"
+#include "common/crypto/kerl/converter.h"
+#include "common/crypto/kerl/kerl.h"
 
 #define RATE 832
 #define CAPACITY (1600 - RATE)
