+++ common/crypto/kerl/kerl.h
@@ -5,10 +5,6 @@
  * Refer to the LICENSE file for licensing information
  */
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 #ifndef __COMMON_KERL_KERL_H_
 #define __COMMON_KERL_KERL_H_
 
@@ -22,6 +18,10 @@ extern "C" {
 #include "common/stdint.h"
 #include "common/trinary/trits.h"
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 typedef struct {
   Keccak_HashInstance keccak;
 } Kerl;
@@ -32,7 +32,8 @@ void kerl_absorb(Kerl* const ctx, trit_t const* trits, size_t const length);
 void kerl_squeeze(Kerl* const ctx, trit_t* trits, size_t const length);
 void kerl_reset(Kerl* const ctx);
 
-#endif
 #ifdef __cplusplus
 }
 #endif
+
+#endif  // __COMMON_KERL_KERL_H_s
