+++ common/crypto/kerl/tests/BUILD
@@ -2,7 +2,7 @@ cc_test(
     name = "test_converter",
     srcs = ["test_converter.c"],
     deps = [
-        "//common/kerl:converter",
+        "//common/crypto/kerl:converter",
         "//common/trinary:trit_tryte",
         "@unity",
     ],
@@ -12,7 +12,7 @@ cc_test(
     name = "test_kerl",
     srcs = ["test_kerl.c"],
     deps = [
-        "//common/kerl",
+        "//common/crypto/kerl",
         "//common/trinary:trit_tryte",
         "@unity",
     ],
