+++ common/model/bundle.h
@@ -8,8 +8,8 @@
 #ifndef __COMMON_MODEL_BUNDLE_H__
 #define __COMMON_MODEL_BUNDLE_H__
 
+#include "common/crypto/kerl/kerl.h"
 #include "common/errors.h"
-#include "common/kerl/kerl.h"
 #include "common/model/transaction.h"
 #include "common/sign/normalize.h"
 #include "common/trinary/flex_trit.h"
