+++ common/sign/v1/iss_kerl.c
@@ -6,7 +6,7 @@
  */
 
 #include "iss_kerl.h"
-#include "common/kerl/kerl.h"
+#include "common/crypto/kerl/kerl.h"
 
 #define HASH_PREFIX kerl
 #define HASH_STATE Kerl
