+++ common/sign/v1/iss_kerl.h
@@ -8,7 +8,7 @@
 #ifndef __ISS_KERL_H_
 #define __ISS_KERL_H_
 
-#include "common/kerl/kerl.h"
+#include "common/crypto/kerl/kerl.h"
 
 #define HASH_PREFIX kerl
 #define HASH_STATE Kerl
