+++ common/sign/v2/iss_kerl.c
@@ -5,7 +5,7 @@
  * Refer to the LICENSE file for licensing information
  */
 
-#include "common/kerl/kerl.h"
+#include "common/crypto/kerl/kerl.h"
 
 #define HASH_PREFIX kerl
 #define HASH_STATE Kerl
